requirements.txt
apache-airflow-providers-celery==3.12.4
apache-airflow-providers-standard==1.4.1
apache-airflow-providers-odbc==4.7.1
apache-airflow-providers-cncf-kubernetes==10.6.2
apache-airflow-providers-amazon[python3-saml]==9.16.0
# Do not upgrade to past 5 until this is resolved https://github.com/apache/airflow/issues/25904
apache-airflow-providers-tableau==4.6.1
apache-airflow-providers-salesforce==5.11.3
apache-airflow-providers-google==16.0.0
apache-airflow-providers-snowflake==6.5.4
dropbox==10.2.0
gspread==3.1.0
pygsheets==2.0.2
pytzdata==2020.1
xlrd==1.2.0
asyncpg==0.30.0
psycopg2-binary==2.9.11
google-analytics-data==0.18.16
sqlalchemy==2.0.44
s3fs